Datetime reformat when reading config file for metadata - Fixes #17 (#18)
* New ISO formatting for datetime items in config file. Fixes #17

Fixes an issue that caused the control panel to have an internal server error when datetime datatypes were passed in as metadata. Implements a method called metadata_as_string that uses the datetime function isoformat() to convert the dates to strings.

* Adding scroll bar to control panel and fixing resizing error from #13

This changes the styling in control-panel.html to not have overlapping elements depending on window sizes. It also implements a scroll bar if the content does not fit on the screen. May not be the best permanent fix but it addresses the main bug.

* Date or datetime now works with control panel - #17

Both date and datetime datatypes are now converted to strings through data_as_string. However, there are still some weaknesses including incompatibility with additional datetime datatypes. This commit also loads the data from the config file into filedata before passing it into metadata. No progress has been made on making the data_as_string function more efficient.

* Compatability with all time types

Ensures that all datetime data types work with the data_as_string function