File reorganization and adding buttons
Reorganized code into separate files. stars.js is similar to the bars.js and creates the experiment controllers and full screen button for the experiment. It also creates and starts StarfieldArena. starfield_arena.js creates and exports a class that is used to create the threejs environment and objects within it. It also has all functions related to animation.